+++ nlppln/commands/ls.py
@@ -9,7 +9,8 @@ from ..utils import cwl_file
 @click.command()
 @click.argument('dir_in', type=click.Path(exists=True))
 @click.option('--recursive/--no-recursive', default=False)
-def command(dir_in, recursive):
+@click.option('--endswith', default=None)
+def command(dir_in, recursive, endswith):
     files_out = []
 
     if recursive:
@@ -21,6 +22,11 @@ def command(dir_in, recursive):
             fi = os.path.join(dir_in, f)
             if os.path.isfile(fi):
                 files_out.append(cwl_file(fi))
+
+    # filter
+    if endswith:
+        files_out = filter(lambda x: x.get('path').endswith(endswith),
+                           files_out)
 
     # order alphabetically on file name
     files_out = sorted(files_out, key=lambda x: x.get('path'))