+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Changed
 
 - Remove erroneous arguments section from `freqs.cwl`
+- Add option to filter on filename to `ls.cwl`
 
 ## 0.3.0